First working menu action: Rescan devices
